Network.Socks5.Lowlevel

Description

Low level types and functions for interacting with a SOCKS server.

Documentation

socksListen

Arguments

:: Socket
The socket to use.

-> IO SocksRequest

For the specified socket, wait for a SOCKS Hello, send a SOCKS Hello response (specifying no authentification method), and wait for a SOCKS request.

Low level types and functions

newtype SocksHello

Type representing initial messages sent by a client with the list of authentification methods supported.

Constructors

SocksHello

Fields

getSocksHelloMethods :: [SocksMethod]

newtype SocksHelloResponse

Type representing initial messages sent by a server in response to Hello, with the server's chosen method of authentication.

Constructors

SocksHelloResponse

Fields

getSocksHelloResponseMethod :: SocksMethod

data SocksRequest

Type representing SOCKS requests.

Constructors

SocksRequest

Fields

requestCommand :: SocksCommand

requestDstAddr :: SocksHostAddress

requestDstPort :: PortNumber

data SocksResponse

Type representing SOCKS responses.

Constructors

SocksResponse

Fields

responseReply :: SocksReply

responseBindAddr :: SocksHostAddress

responseBindPort :: PortNumber

class Command a where

A type class for types that can yield a SOCKS request or, optionally, be obtained from a SOCKS request.

Methods

toRequest :: a -> SocksRequest

fromRequest :: SocksRequest -> Maybe a

newtype Connect

Type representing connect commands.

Constructors

Connect SocksAddress

establish

Arguments

:: SocksVersion
The SOCKS protocol version to use.

-> Socket
The socket to use.

-> [SocksMethod]
A list of the authentification methods supported.

-> IO SocksMethod

Establish a connection with a SOCKS server.

rpc_

Arguments

:: Command a

=> Socket
The socket to use.

-> a
The command.

-> IO (SocksHostAddress, PortNumber)

As for rpc but throws an exception if it does not succeed.

connectIPV4

Arguments

:: Socket
The socket to use.

-> HostAddress
The host address.

-> PortNumber
The port number to use.

-> IO (HostAddress, PortNumber)

Connect using IPv4.

connectIPV6

Arguments

:: Socket
The socket to use.

-> HostAddress6
The host address.

-> PortNumber
The port number to use.

-> IO (HostAddress6, PortNumber)

Connect using IPv6.

connectDomainName

Arguments

:: Socket
The socket to use.

-> String
Destination FQDN. Should comprise only ASCII characters, otherwise unexpected behaviour will ensure. For FQDN including other Unicode code points, Punycode encoding should be used.

-> PortNumber
The port number to use.

-> IO (SocksHostAddress, PortNumber)

Connect using a fully-qualified domain name (FQDN).

rpc

Arguments

:: Command a

=> Socket
The socket to use.

-> a
The command.

-> IO (Either SocksError (SocksHostAddress, PortNumber))

Try to execute the specified command with the specified socket.

sendSerialized

Arguments

:: Serialize a

=> Socket
The socket to use.

-> a
The data.

-> IO ()

Send data to the specified socket.

waitSerialized

Arguments

:: Serialize a

=> Socket
The socket to use.

-> IO a

Wait for data from the specified socket.