Network.Socks5.Types

Description

Documentation

data SocksVersion

Type representing SOCKS protocol versions.

Constructors

SocksVer5
SOCKS Protocol Version 5. The only version implemented by the library.

data SocksCommand

Type representing commands that can be sent or received under the SOCKS protocol.

Constructors

SocksCommandConnect
The CONNECT request.

SocksCommandBind
The BIND request. Not implemented by the library.

SocksCommandUdpAssociate
The UDP ASSOCIATE request. Not implemented by the library.

SocksCommandOther !Word8
Other requests. None are specified by the SOCKS Protocol Version 5.

data SocksMethod

Type representing authentication methods available under the SOCKS protocol.

Only SocksMethodNone is effectively implemented, but other values are enumerated for completeness.

Constructors

SocksMethodNone
NO AUTHENTICATION REQUIRED.

SocksMethodGSSAPI
GSSAPI.

SocksMethodUsernamePassword
USERNAME/PASSWORD.

SocksMethodOther !Word8
IANA ASSIGNED or RESERVED FOR PRIVATE METHODS.

SocksMethodNotAcceptable
NO ACCEPTABLE METHODS.

data SocksHostAddress

Type representing host addresses under the SOCKS protocol.

Constructors

SocksAddrIPV4 !HostAddress
A version-4 IP address.

SocksAddrDomainName !SocksFQDN
A fully-qualified domain name (FQDN).

SocksAddrIPV6 !HostAddress6
A version-6 IP address.

data SocksAddress

Type representing socket addresses under the SOCKS protocol.

Constructors

SocksAddress !SocksHostAddress !PortNumber

data SocksReply

Type representing replies under the SOCKS protocol.

Constructors

SocksReplySuccess
The server reports that the request succeeded.

SocksReplyError SocksError
The server reports that the request did not succeed.

data SocksVersionNotSupported

Type representing exceptions.

Constructors

SocksVersionNotSupported
The SOCKS protocol version is not supported. This library only implements SOCKS Protocol Version 5.

data SocksError

Type representing SOCKS errors that can be part of a SOCKS reply.

Constructors

SocksErrorGeneralServerFailure
General SOCKS server failure.

SocksErrorConnectionNotAllowedByRule
Connection not allowed by ruleset.

SocksErrorNetworkUnreachable
Network unreachable.

SocksErrorHostUnreachable
Host unreachable.

SocksErrorConnectionRefused
Connection refused.

SocksErrorTTLExpired
TTL expired.

SocksErrorCommandNotSupported
Command not supported.

SocksErrorAddrTypeNotSupported
Address type not supported.

SocksErrorOther Word8
Other error. Unassigned in SOCKS Protocol Version 5.

type SocksFQDN = ByteString

Type synonym representing fully-qualified domain names (FQDN). The SOCKS Protocol Version 5 does not specify an encoding for a FQDN other than there is no terminating NUL octet (byte).

This library's API assumes that FQDN values comprise only ASCII characters. Domain names that include other Unicode code points should be Punycode encoded.